Split big file into two smaller units
