+++ link-grammar/dict-common/dict-locale.h
@@ -1,0 +1,13 @@
+
+/**
+ * Dictionary locale management.
+ */
+#ifndef _DICT_LOCALE_H_
+#define _DICT_LOCALE_H_
+
+#include "link-includes.h"
+
+void dictionary_setup_locale(Dictionary dict);
+bool dictionary_setup_defines(Dictionary dict);
+
+#endif /* _DICT_LOCALE_H_ */
