+++ link-grammar/dict-sql/read-sql.c
@@ -23,9 +23,10 @@
 
 #include "api-structures.h"
 #include "connectors.h"
+#include "dict-common/dict-affix-impl.h"
 #include "dict-common/dict-api.h"
 #include "dict-common/dict-common.h"
-#include "dict-common/dict-impl.h"
+#include "dict-common/dict-locale.h"
 #include "dict-common/dict-structures.h"
 #include "dict-common/dict-utils.h"      // patch_subscript()
 #include "dict-common/file-utils.h"
